we try to make decentralized social media application.
NitroGenZ is a combination of two words. “Nitro” means Nitrogen, which denotes Stability. “GenZ” which denotes Generation Z. The idea behind this name is that a Dapp made by generation Z for generation Z.Our mission is to preserve data from social media giants. This includes providing users with full protection for their data. Since we do not use the database to store and process user data, we are unlike many others. Besides, this gives the users extra protection against cybercriminals.